requirements.txt
cryptography>=38.0.1
pycrypto>=2.6.1
pyyaml>=6.0
keyring>=23.9.3